fixed issue with oauth login

+++ db/migrate/20190102175336_add.rb
@@ -1,5 +1,5 @@
 class Add < ActiveRecord::Migration[5.2]
   def change
-    add_column :users, :uid, :integer
+    add_column :users, :uid, :numeric
   end
 end
